requirements.txt
Django
django-filter
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-decouple
Pillow
psycopg2-binary
drf-yasg
celery
redis
requests
